fix bunch of things

+++ src/Services/DAO.py
@@ -2,8 +2,8 @@ import os
 import pandas as pd
 from datetime import datetime
 
-CASH_POSITION_FILE = os.getenv("CASH_POSITION_FILE")
-STOCK_POSITION_FILE = os.getenv("STOCK_POSITION_FILE")
+CASH_POSITION_FILE = "CASH_POSITION_FILE"
+STOCK_POSITION_FILE = "STOCK_POSITION_FILE"
 
 def save_cash_position(cash_balance):
     data = {
@@ -11,11 +11,12 @@ def save_cash_position(cash_balance):
         'cash_balance': [cash_balance]
     }
     df = pd.DataFrame(data)
-    df.to_csv(CASH_POSITION_FILE, mode='a', index=False, header=False)
+    df.to_csv(CASH_POSITION_FILE, mode='w', index=False, header=True)
 
 def get_remaining_cash():
     try:
         cash_data = pd.read_csv(CASH_POSITION_FILE)
+        print(cash_data['cash_balance'].iloc[-1])
         return cash_data['cash_balance'].iloc[-1]
     except FileNotFoundError:
         return 0  
@@ -28,12 +29,16 @@ def save_stock_position(symbol, quantity, price):
         'buying_price': [price]
     }
     df = pd.DataFrame(data)
-    df.to_csv(STOCK_POSITION_FILE, mode='a', index=False, header=False)
 
- def update_stock_position(trading_date, symbol, new_quantity, buying_price):
+    if not os.path.exists(STOCK_POSITION_FILE):
+        df.to_csv(STOCK_POSITION_FILE, mode='w', index=False, header=True)
+    else:
+        df.to_csv(STOCK_POSITION_FILE, mode='a', index=False, header=False)
+
+def update_stock_position(trading_date, symbol, new_quantity, buying_price):
     df = pd.read_csv(STOCK_POSITION_FILE)
 
-    # Update the specific row where symbol and trading_date match
+#     # Update the specific row where symbol and trading_date match
     df.loc[(df['trading_date'] == trading_date) & (df['symbol'] == symbol), 'quantity'] = new_quantity
     df.to_csv(STOCK_POSITION_FILE, index=False)
 
@@ -54,7 +59,7 @@ def get_all_stocks():
         return stock_data  
     except FileNotFoundError:
         return pd.DataFrame()
-    
+ 
 def delete_stock_position(trading_date, symbol):
     df = pd.read_csv(STOCK_POSITION_FILE)
 
@@ -62,20 +67,17 @@ def delete_stock_position(trading_date, symbol):
     df = df[~((df['trading_date'] == trading_date) & (df['symbol'] == symbol))]
     df.to_csv(STOCK_POSITION_FILE, index=False)
 
-def init_cash_position(self):
+def init_cash_position():
     if not os.path.exists(CASH_POSITION_FILE):
         data = {'date': [pd.Timestamp.now().strftime('%Y-%m-%d')], 'cash_balance': [0.00]}  # Initial balance $0
         df = pd.DataFrame(data)
         df.to_csv(CASH_POSITION_FILE, index=False)
-        print(f"Initialized {CASH_POSITION_FILE} with $10,000 balance.")
+        print(f"Initialized {CASH_POSITION_FILE} with $0 balance.")
     else:
         print(f"{CASH_POSITION_FILE} already exists.")
 
-def init_stock_position(self):
-    if not os.path.exists(STOCK_POSITION_FILE):
-        data = {'trading_date': [], 'symbol': [], 'quantity': [], 'buying_price': []}
-        df = pd.DataFrame(data)
-        df.to_csv(STOCK_POSITION_FILE, index=False)
-        print(f"Initialized {STOCK_POSITION_FILE} as an empty file.")
-    else:
-        print(f"{STOCK_POSITION_FILE} already exists.")
+def init_stock_position():
+    data = {'trading_date': [], 'symbol': [], 'quantity': [], 'buying_price': []}
+    df = pd.DataFrame(data)
+    df.to_csv(STOCK_POSITION_FILE, index=False)
+    print(f"Initialized {STOCK_POSITION_FILE} as an empty file.")
